requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# File Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
python-multipart==0.0.6
Pillow==10.2.0

# OCR (Optional but recommended)
pytesseract==0.3.10
pdf2image==1.17.0

# AI/ML - Hugging Face
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99
accelerate==0.25.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.26.0